src/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
configparser==3.7.4
cryptography==2.7
Django==2.2.3
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
mysqlclient==1.4.1
protobuf==3.8.0
psycopg2==2.8.3
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
